✨ feat: installation (-i)
Handling the -i flag: The script now checks if the -i flag is provided and sets the INSTALL variable to true if it is. The version number is then taken from the second argument. Installation section: When INSTALL is true, the script proceeds with cloning, building, and installing the specified PHP version. Symbolic links creation: After installation (if needed) and checking the PHP binaries, the script creates symbolic links as before.

To activate an already installed PHP version: phpenv 81 To install and activate a PHP version: phpenv -i 83